refactor: Расчёт общей стоимости выделен в отдельный компонент
